refactor: Complete architectural domain separation (Phases 2-5)

+++ TelegramGroupsAdmin.ContentDetection/Models/ContentCheckResult.cs
@@ -33,7 +33,7 @@ public record ContentCheckResult
     /// <summary>
     /// Total processing time for all checks in milliseconds
     /// </summary>
-    public long TotalProcessingTimeMs { get; init; } = 0;
+    public double TotalProcessingTimeMs { get; init; } = 0;
 
     /// <summary>
     /// All extra message IDs to delete (combined from all checks)
